Fixed TyperError in function 'ricker' for tuning_wedge_v2.py and tuning_prestack_v2.py. 'np.linspace' command was failing when creating variable 't' because the number of samples to generate should be an Integer, but the provided value was Float

+++ 1412_Tuning_and_AVO/tuning_wedge_v2.py
@@ -136,7 +136,7 @@ def ricker(cfreq, phase, dt, wvlt_length):
     
     t = np.arange(t_min, t_max, dt)
     
-    t = np.linspace(-wvlt_length/2, (wvlt_length-dt)/2, wvlt_length/dt)
+    t = np.linspace(-wvlt_length/2, (wvlt_length-dt)/2, int(wvlt_length/dt))
     wvlt = (1.0 - 2.0*(np.pi**2)*(cfreq**2)*(t**2)) * np.exp(-(np.pi**2)*(cfreq**2)*(t**2))
     
     if phase != 0:
